[Refactor] 장비 네트워크 엔트리 타입 파일 분리
