+++ RAW_unity/Assets/_RAW/Scripts/Network/Equipment/NetworkEquipmentEntry.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: bd9f9cf8aaee247b197d759861a5fe72
